pyproject.toml
[tool.poetry]
name = "hft-scalping-bot"
version = "0.1.0"
description = "High-frequency trading bot with ML-powered market prediction"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "hft_scalping_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = {version = "^1.21.0", source = "pypi"}
pandas = {version = "^1.3.0", source = "pypi"}
scikit-learn = "^1.0.0"
torch = {version = "^1.10.0", source = "pypi"}
ccxt = "^2.0.0"
aiohttp = "^3.8.0"
websockets = "^10.0"
pyyaml = "^6.0"
python-dotenv = "^0.19.0"
structlog = "^21.1.0"
prometheus-client = "^0.12.0"
mlflow = "^2.1.0"
pydantic = "^2.0"
pyarrow = {version = "^11.0.0", source = "pypi"}
lightgbm = "^3.3.0"
boto3 = "^1.26.0"
azure-keyvault-secrets = "^4.6.0"
azure-identity = "^1.12.0"
google-cloud-secret-manager = "^2.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.17.0"
pytest-cov = "^3.0.0"
black = "^22.0.0"
mypy = "^0.900"
flake8 = "^4.0.0"

[[tool.poetry.source]]
name = "pypi"
url = "https://pypi.org/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
warn_unreachable = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=hft_scalping_bot --cov-report=term-missing"